+++ PseudoAbsences/src/pseudoabsencemask.jl
@@ -1,4 +1,4 @@
-function _layer_works_for_pseudoabsence(layer::SDMLayer{T}) where {T <: Bool}
+function _layer_works_for_pseudoabsence(layer::SDMLayer{T}) where {T<:Bool}
     iszero(sum(layer)) && throw(ArgumentError("The presences layer is empty"))
     return nothing
 end
@@ -19,8 +19,8 @@ end
 @testitem "We can use RandomSelection" begin
     using PseudoAbsences.SimpleSDMLayers
     n = SDMLayer(zeros(Bool, 10, 10))
-    n.grid[1]  = true
-    n.grid[end]  = true
+    n.grid[1] = true
+    n.grid[end] = true
     p = pseudoabsencemask(RandomSelection, n)
     @test all(p.grid[2:(end-1)])
 end
@@ -53,17 +53,17 @@ end
 @testitem "We can use SurfaceRangeEnvelope" begin
     using PseudoAbsences.SimpleSDMLayers
     n = SDMLayer(zeros(Bool, 10, 10))
-    n.grid[2,2] = true
-    n.grid[9,9] = true
+    n.grid[2, 2] = true
+    n.grid[9, 9] = true
     p = pseudoabsencemask(SurfaceRangeEnvelope, n)
     for i in axes(n.grid, 1)
         for j in axes(n.grid, 2)
             if (2 <= i <= 9) & (2 <= j <= 9)
                 # Test the surface range
-                @test p.grid[i,j] == !n.grid[i,j]
+                @test p.grid[i, j] == !n.grid[i, j]
             else
                 # Test the exterior borders (all false)
-                @test p.grid[i,j]  == false
+                @test p.grid[i, j] == false
             end
         end
     end
@@ -79,7 +79,7 @@ distances, it may be a very good idea to flatten this layer using `log` or an
 exponent. The `f` function is used to determine which distance is reported
 (`minimum` by default, but can be any other relevant function).
 """
-function pseudoabsencemask(::Type{DistanceToEvent}, presences::SDMLayer{Bool}; f = minimum)
+function pseudoabsencemask(::Type{DistanceToEvent}, presences::SDMLayer{Bool}; f=minimum)
     _layer_works_for_pseudoabsence(presences)
     presence_only = nodata(presences, false)
     background = zeros(presences, Float64)
@@ -87,7 +87,7 @@ function pseudoabsencemask(::Type{DistanceToEvent}, presences::SDMLayer{Bool}; f
     prj = SimpleSDMLayers.Proj.Transformation(
         presences.crs,
         "+proj=longlat +datum=WGS84 +no_defs";
-        always_xy = true,
+        always_xy=true,
     )
     E, N = eastings(presences), northings(presences)
 
@@ -113,25 +113,73 @@ function pseudoabsencemask(::Type{DistanceToEvent}, presences::SDMLayer{Bool}; f
     return background
 end
 
+function _zero_occ!(bg, presences)
+    for i in keys(nodata(presences, false))
+        bg[i] = false
+    end
+    return bg
+end
+
 """
     pseudoabsencemask(::Type{WithinRadius}, presences::SDMLayer{Bool}; distance::Number = 100.0, )
 
 Generates a mask for pseudo-absences where pseudo-absences can be within a
 `distance` (in kilometers) of the original observation. Internally, this uses
-`DistanceToEvent`.
+`DistanceToEvent`. All other keyword arguments are passed to
+`pseudoabsencemask`.
 """
 function pseudoabsencemask(
     ::Type{WithinRadius},
     presences::SDMLayer{Bool};
-    distance::Number = 100.0,
+    distance::Number=100.0,
+    kwargs...
 )
     _layer_works_for_pseudoabsence(presences)
-    bg = pseudoabsencemask(DistanceToEvent, presences; f = minimum)
+    bg = pseudoabsencemask(DistanceToEvent, presences; kwargs...)
     background = bg .<= distance
-    for i in keys(nodata(presences, false))
-        background[i] = false
-    end
-    return background
+    return _zero_occ!(background, presences)
+end
+
+"""
+    pseudoabsencemask(::Type{WithoutRadius}, presences::SDMLayer{Bool}; distance::Number = 100.0, )
+
+Generates a mask for pseudo-absences where pseudo-absences can be outside of a
+`distance` (in kilometers) of the original observation. Internally, this uses
+`DistanceToEvent`. All other keyword arguments are passed to
+`pseudoabsencemask`.
+"""
+function pseudoabsencemask(
+    ::Type{WithoutRadius},
+    presences::SDMLayer{Bool};
+    distance::Number=100.0,
+    kwargs...
+)
+    _layer_works_for_pseudoabsence(presences)
+    bg = pseudoabsencemask(DistanceToEvent, presences; kwargs...)
+    background = bg .>= distance
+    return _zero_occ!(background, presences)
+end
+
+
+"""
+    pseudoabsencemask(::Type{BetweenRadius}, presences::SDMLayer{Bool}; closer::Number = 10.0, further::Number=100.0)
+
+Generates a mask for pseudo-absences where pseudo-absences can be in a band
+given by two distances (in kilometers) of the original observation. Internally,
+this uses `DistanceToEvent`. All other keyword arguments are passed to
+`pseudoabsencemask`.
+"""
+function pseudoabsencemask(
+    ::Type{BetweenRadius},
+    presences::SDMLayer{Bool};
+    closer::Number=10.0,
+    further::Number=100.0,
+    kwargs...
+)
+    _layer_works_for_pseudoabsence(presences)
+    bg = pseudoabsencemask(DistanceToEvent, presences; kwargs...)
+    background = closer .<= bg .<= further
+    return _zero_occ!(background, presences)
 end
 
 """
@@ -147,7 +195,7 @@ to the projection / position.
 When `absolute` is `false` (the default), the distance is measured as the
 absolute difference in coordinates instead of using the Euclidean distance.
 """
-function pseudoabsencemask(::Type{DegreesToEvent}, presences::SDMLayer{Bool}; f = minimum, absolute::Bool = false)
+function pseudoabsencemask(::Type{DegreesToEvent}, presences::SDMLayer{Bool}; f=minimum, absolute::Bool=false)
     _layer_works_for_pseudoabsence(presences)
     presence_only = nodata(presences, false)
     background = zeros(presences, Float64)
@@ -155,14 +203,14 @@ function pseudoabsencemask(::Type{DegreesToEvent}, presences::SDMLayer{Bool}; f 
     prj = SimpleSDMLayers.Proj.Transformation(
         presences.crs,
         "+proj=longlat +datum=WGS84 +no_defs";
-        always_xy = true,
+        always_xy=true,
     )
     E, N = eastings(presences), northings(presences)
 
     points = [prj(E[i.I[2]], N[i.I[1]]) for i in keys(presence_only)]
 
     # Distance function used for degrees
-    distfunc(pk, ko) = absolute ? sqrt(sum((pk .- ko).^2.0)) : sum(abs.(pk .- ko))
+    distfunc(pk, ko) = absolute ? sqrt(sum((pk .- ko) .^ 2.0)) : sum(abs.(pk .- ko))
 
     # Prepare for thread-safe parallelism
     bg = keys(background)
@@ -181,4 +229,66 @@ function pseudoabsencemask(::Type{DegreesToEvent}, presences::SDMLayer{Bool}; f 
     fetch.(tasks)
 
     return background
+end
+
+
+
+"""
+    pseudoabsencemask(::Type{WithinDegrees}, presences::SDMLayer{Bool}; distance::Number = 3.0, )
+
+Generates a mask for pseudo-absences where pseudo-absences can be within a
+`distance` (in degrees) of the original observation. Internally, this uses
+`DegreesToEvent`. All other keyword arguments are passed to `pseudoabsencemask`.
+"""
+function pseudoabsencemask(
+    ::Type{WithinDegrees},
+    presences::SDMLayer{Bool};
+    distance::Number=3.0,
+    kwargs...
+)
+    _layer_works_for_pseudoabsence(presences)
+    bg = pseudoabsencemask(DegreesToEvent, presences; f=kwargs...)
+    background = bg .<= distance
+    return _zero_occ!(background, presences)
+end
+
+"""
+    pseudoabsencemask(::Type{WithoutDegrees}, presences::SDMLayer{Bool}; distance::Number = 3.0, )
+
+Generates a mask for pseudo-absences where pseudo-absences can be outside of a
+`distance` (in degrees) of the original observation. Internally, this uses
+`DegreesToEvent`. All other keyword arguments are passed to `pseudoabsencemask`.
+"""
+function pseudoabsencemask(
+    ::Type{WithoutDegrees},
+    presences::SDMLayer{Bool};
+    distance::Number=3.0,
+    kwargs...
+)
+    _layer_works_for_pseudoabsence(presences)
+    bg = pseudoabsencemask(DegreesToEvent, presences; kwargs...)
+    background = bg .>= distance
+    return _zero_occ!(background, presences)
+end
+
+
+"""
+    pseudoabsencemask(::Type{BetweenDegrees}, presences::SDMLayer{Bool}; closer::Number = 1.0, further::Number=3.0)
+
+Generates a mask for pseudo-absences where pseudo-absences can be in a band
+given by two distances (in degrees) of the original observation. Internally,
+this uses `DegreesToEvent`. All other keyword arguments are passed to
+`pseudoabsencemask`.
+"""
+function pseudoabsencemask(
+    ::Type{BetweenDegrees},
+    presences::SDMLayer{Bool};
+    closer::Number=10.0,
+    further::Number=100.0,
+    kwargs...
+)
+    _layer_works_for_pseudoabsence(presences)
+    bg = pseudoabsencemask(DegreeToEvent, presences; kwargs...)
+    background = closer .<= bg .<= further
+    return _zero_occ!(background, presences)
 end